feat(tunnel): 添加设备隧道功能支持frp穿透
- 在环境变量配置中增加FRPS相关参数设置
- 创建数据库迁移脚本添加device_tunnels表结构
- 前端界面集成Cable图标并添加隧道管理标签页
- 实现隧道启用、禁用、更新和配置生成功能
- 添加隧道相关的API端点和数据模型定义
- 集成frps服务到Docker容器启动流程中
- 添加隧道操作的审计日志记录功能
- 实现隧道配置文件自动生成和下载功能
- 添加隧道令牌轮换和状态管理功能
- 集成端到端测试验证隧道工作流程

+++ app/models/audit_log.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import JSON, BigInteger, DateTime, Index, String, func
+from sqlalchemy import JSON, BigInteger, DateTime, Index, Integer, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db.base import Base
@@ -15,7 +15,7 @@ class AuditLog(Base):
         Index("idx_audit_device", "device_id"),
     )
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int | None] = mapped_column(BigInteger)
     action: Mapped[str] = mapped_column(String(128), nullable=False)
     device_id: Mapped[str | None] = mapped_column(String(128))
